backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai==1.12.0
pydantic==2.6.0
chromadb==0.4.22
tiktoken==0.5.2